Fix handling of Tx when opening a pack

+++ pages/Play/index.js
@@ -348,7 +348,7 @@ const Play = (props) => {
 
     if (claimData && claimData.winning_placements.length > 0) {
       totalPrize = claimData.winning_placements.reduce((total,   num) => {
-        let acc = parseInt(total) + parseInt(num.prize);
+        let acc = parseFloat(total) + parseFloat(num.prize);
         return acc;
       }, 0);
     }
